+++ gmail.js
@@ -1,60 +1,50 @@
-class MainScript {
+class gmailMainScript {
   constructor() {
-    this.configureApp()
-    this.btn = null
-    this.token = null
-    this.fullNameElement = document.querySelector("input[name='name']")
-    this.emailElement = document.querySelector("input[name='email']")
-    this.phoneElement = document.querySelector("input[name='phone']")
-    this.companyElement = document.querySelector("input[name='org']")
-    this.linkedinElement = document.querySelector(
-      "input[name='urls[LinkedIn]']"
-    )
+    this.configureApp();
+    this.btn = null;
+    this.token = null;
+    this.fullNameElement = document.querySelector("input[name='name']");
+    this.emailElement = document.querySelector("input[name='email']");
+    this.phoneElement = document.querySelector("input[name='phone']");
+    this.companyElement = document.querySelector("input[name='org']");
+    this.linkedinElement = document.querySelector("input[name='urls[LinkedIn]']");
   }
 
-  fillnameEmailCompanyLinkedIn(
-    fullName,
-    email,
-    phone,
-    currentCompany,
-    linkedin
-  ) {
-    this.fullNameElement.value = fullName && fullName
-    this.emailElement.value = email && email
-    this.phoneElement.value = phone && phone
-    this.companyElement.value = currentCompany && currentCompany
-    this.linkedinElement.value = linkedin
+  fillnameEmailCompanyLinkedIn(fullName, email, phone, currentCompany, linkedin) {
+    this.fullNameElement.value = fullName && fullName;
+    this.emailElement.value = email && email;
+    this.phoneElement.value = phone && phone;
+    this.companyElement.value = currentCompany && currentCompany;
+    this.linkedinElement.value = linkedin;
   }
+
   async uploadResume(resume_url, filename) {
-    // website.value = data.data.website && data.data.website;
     const getPDF = await fetch(resume_url, {
       method: "GET",
       mode: "cors",
-    })
+    });
 
-    const pdfBlob = await getPDF.blob()
+    const pdfBlob = await getPDF.blob();
 
-    //Gmail needs a .pdf appended to the filename for it to work
     const myFile = new File([pdfBlob], filename + ".pdf", {
       type: "application/pdf",
       lastModified: new Date(),
-    })
+    });
 
-    const dataTransfer = new DataTransfer()
-    dataTransfer.items.add(myFile)
-    const dec = document.querySelector('input[type="file"][name="Filedata"]')
-    dec.files = dataTransfer.files
+    const dataTransfer = new DataTransfer();
+    dataTransfer.items.add(myFile);
+    const dec = document.querySelector('input[type="file"][name="Filedata"]');
+    dec.files = dataTransfer.files;
 
-    const event = new Event("change")
-
-    dec.dispatchEvent(event)
+    const event = new Event("change");
+    dec.dispatchEvent(event);
   }
 
   async handlePopUpbuttonClicked() {
-    const session = (await chrome.storage.sync.get("session"))["session"]
+    const session = (await chrome.storage.sync.get("session"))["session"];
 
     if (session) {
-      console.log(session)
+      console.log(session);
       const response = await fetch("https://instantapply.co/api/getUser", {
         method: "POST",
         mode: "cors",
@@ -64,17 +54,15 @@ class MainScript {
         headers: {
           "Content-Type": "application/json",
         },
-      })
+      });
 
       if (response.status === 401)
-        window.open("https://instantapply.co", "_blank")
+        window.open("https://instantapply.co", "_blank");
 
       if (response.status === 200) {
-
-        // Click on the compose button
-        const clickCompose = document.querySelector(".T-I.T-I-KE.L3")
-        clickCompose.click()
-        this.data = await response.json()
+        const clickCompose = document.querySelector(".T-I.T-I-KE.L3");
+        clickCompose.click();
+        this.data = await response.json();
         const {
           data: {
             lastname,
@@ -86,60 +74,50 @@ class MainScript {
             filename,
             currentCompany,
           },
-        } = this.data
+        } = this.data;
         this.fillnameEmailCompanyLinkedIn(
           `${firstname} ${lastname}`,
           resume_email,
           phone,
           currentCompany,
           linkedin
-        )
-        this.uploadResume(resume_url, filename)
+        );
+        this.uploadResume(resume_url, filename);
       }
-    } else window.open("https://instantapply.co", "_blank")
+    } else window.open("https://instantapply.co", "_blank");
   }
 
   addPopUpButtonToPage() {
-    var btn = document.createElement("div")
-    btn.classList.add("action_button")
-    var img = document.createElement("img")
-    img.src = "https://instantapply.co/assets/images/instantapply-logo.svg"
-    img.style.width = "50px"
-    img.style.height = "50px"
-    btn.appendChild(img)
+    var btn = document.createElement("div");
+    btn.classList.add("action_button");
+    var img = document.createElement("img");
+    img.src = "https://instantapply.co/assets/images/instantapply-logo.svg";
+    img.style.width = "50px";
+    img.style.height = "50px";
+    btn.appendChild(img);
     //styling the button
-    img.style.position = "fixed"
-    img.style.top = "10%"
-    img.style.right = "7%"
-    img.style.borderRadius = "8px"
-    img.style.backgroundColor = "#ede2ff"
-    img.style.padding = "10px"
-    img.style.zIndex = "999"
-    // btn.style.height = "50px";
-    // btn.style.width = "50px";
-    // btn.style.color = "white";
-    // btn.style.fontSize = "22px";
-    // btn.style.fontWeight = "bold";
-    // btn.style.backgroundColor = "#999";
-    // btn.style.borderRadius = "50%";
-    // btn.style.display = "flex";
-    // btn.style.alignItems = "center";
-    // btn.style.justifyContent = "center";
-    btn.style.cursor = "pointer"
-    document.body.appendChild(btn)
-    this.btn = btn
-    this.configurePopUpButton()
+    img.style.position = "fixed";
+    img.style.top = "10%";
+    img.style.right = "7%";
+    img.style.borderRadius = "8px";
+    img.style.backgroundColor = "#ede2ff";
+    img.style.padding = "10px";
+    img.style.zIndex = "999";
+    btn.style.cursor = "pointer";
+    document.body.appendChild(btn);
+    this.btn = btn;
+    this.configurePopUpButton();
   }
 
-  //add event listener to popup button
   configurePopUpButton() {
-    this.btn.addEventListener("click", this.handlePopUpbuttonClicked.bind(this))
+    this.btn.addEventListener("click", this.handlePopUpbuttonClicked.bind(this));
   }
 
-  //add event listener for when page loads
   configureApp() {
-    window.onload = this.addPopUpButtonToPage()
+    window.addEventListener("load", () => {
+      this.addPopUpButtonToPage();
+    });
   }
 }
 
-new MainScript();
+new gmailMainScript();
