+++ package.json
@@ -4,8 +4,7 @@
   "description": "This is a chrome Extension for easy autofill of Job Applications",
   "main": "background.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "node build.js"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
@@ -18,7 +17,6 @@
   },
   "homepage": "https://github.com/kingebere/instant-apply#readme",
   "devDependencies": {
-    "google-closure-compiler": "^20230502.0.0",
     "terser-webpack-plugin": "^5.3.9",
     "webpack": "^5.87.0",
     "webpack-cli": "^5.1.4"
