+++ submitgreenhouse.js
@@ -1,4 +1,4 @@
-class Main {
+class greenhouseMain {
 	constructor() {
 		//get previous url
 		this.previousUrl = document.referrer;
@@ -40,4 +40,4 @@ class Main {
 	}
 }
 
-new Main();
+new greenhouseMain();
